Rename classes to avoid confusion.
